refactor: moved InfoMessage and ErrorMessage under http package
