+++ event-log/src/test/scala/io/renku/eventlog/MicroserviceRoutesSpec.scala
@@ -20,13 +20,14 @@ package io.renku.eventlog
 
 import cats.effect.{Clock, IO}
 import cats.syntax.all._
-import ch.datascience.controllers.ErrorMessage.ErrorMessage
-import ch.datascience.controllers.InfoMessage.InfoMessage
-import ch.datascience.controllers.{ErrorMessage, InfoMessage}
+import ch.datascience.http.ErrorMessage.ErrorMessage
+import ch.datascience.http.InfoMessage.InfoMessage
+import ch.datascience.http.InfoMessage
 import ch.datascience.generators.Generators.Implicits._
 import ch.datascience.graph.model.EventsGenerators.compoundEventIds
 import ch.datascience.graph.model.GraphModelGenerators.projectIds
 import ch.datascience.graph.model.projects
+import ch.datascience.http.{ErrorMessage, InfoMessage}
 import ch.datascience.http.server.EndpointTester._
 import ch.datascience.interpreters.TestRoutesMetrics
 import ch.datascience.metrics.LabeledGauge
