+++ event-log/src/test/scala/io/renku/eventlog/eventspatching/EventsPatchingEndpointSpec.scala
@@ -20,9 +20,9 @@ package io.renku.eventlog.eventspatching
 
 import cats.effect.IO
 import cats.syntax.all._
-import ch.datascience.controllers.ErrorMessage.ErrorMessage
-import ch.datascience.controllers.InfoMessage._
-import ch.datascience.controllers.{ErrorMessage, InfoMessage}
+import ch.datascience.http.ErrorMessage.ErrorMessage
+import ch.datascience.http.InfoMessage._
+import ch.datascience.http.InfoMessage
 import ch.datascience.generators.Generators.Implicits._
 import ch.datascience.generators.Generators._
 import ch.datascience.graph.model.events.EventStatus
@@ -42,6 +42,7 @@ import org.scalamock.scalatest.MockFactory
 import org.scalatest.matchers.should
 import org.scalatest.wordspec.AnyWordSpec
 import ch.datascience.graph.model.EventsGenerators._
+import ch.datascience.http.{ErrorMessage, InfoMessage}
 
 class EventsPatchingEndpointSpec extends AnyWordSpec with MockFactory with should.Matchers {
