+++ graph-commons/src/test/scala/ch/datascience/http/server/security/AuthenticationSpec.scala
@@ -21,7 +21,7 @@ package ch.datascience.http.server.security
 import cats.data.Kleisli
 import cats.effect.IO
 import cats.syntax.all._
-import ch.datascience.controllers.ErrorMessage._
+import ch.datascience.http.ErrorMessage._
 import ch.datascience.generators.CommonGraphGenerators._
 import ch.datascience.generators.Generators.Implicits._
 import ch.datascience.graph.model.GraphModelGenerators._
