+++ webhook-service/src/main/scala/ch/datascience/webhookservice/eventprocessing/HookEventEndpoint.scala
@@ -23,10 +23,11 @@ import cats.effect._
 import cats.syntax.all._
 import ch.datascience.config.GitLab
 import ch.datascience.control.Throttler
-import ch.datascience.controllers.ErrorMessage._
-import ch.datascience.controllers.{ErrorMessage, InfoMessage}
+import ch.datascience.http.ErrorMessage._
+import ch.datascience.http.InfoMessage
 import ch.datascience.graph.model.events._
 import ch.datascience.graph.model.projects.{Id, Path}
+import ch.datascience.http.{ErrorMessage, InfoMessage}
 import ch.datascience.http.client.RestClientError.UnauthorizedException
 import ch.datascience.logging.ExecutionTimeRecorder
 import ch.datascience.webhookservice.crypto.HookTokenCrypto
