+++ webhook-service/src/test/scala/ch/datascience/webhookservice/eventprocessing/ProcessingStatusEndpointSpec.scala
@@ -22,12 +22,13 @@ import ProcessingStatusGenerator._
 import cats.MonadError
 import cats.data.OptionT
 import cats.effect.IO
-import ch.datascience.controllers.InfoMessage._
-import ch.datascience.controllers.{ErrorMessage, InfoMessage}
+import ch.datascience.http.InfoMessage._
+import ch.datascience.http.InfoMessage
 import ch.datascience.generators.Generators.Implicits._
 import ch.datascience.generators.Generators.exceptions
 import ch.datascience.graph.model.GraphModelGenerators.projectIds
 import ch.datascience.graph.model.projects.Id
+import ch.datascience.http.{ErrorMessage, InfoMessage}
 import ch.datascience.http.client.AccessToken
 import ch.datascience.http.server.EndpointTester._
 import ch.datascience.interpreters.TestLogger
